Refactor exercise routes to use page components
Extract HTML and UI logic from route files into dedicated page components following the project's architectural guidelines in CLAUDE.md. Route files now only contain loader/action functions and minimal default exports.

Changes:
- Create EditExercisePage component for edit route
- Create NewExercisePage component for new route
- Create ExerciseDetailPage component for detail route (includes list styling fix)
- Create ExercisesListPage component for index route
- Update all exercise routes to delegate to page components
- Fix TypeScript imports to use correct types from ~/types
